requirements.txt
azure-functions
pandas
plotly
dash
#gunicorn (necessary?)
#Flask (already dependency of dash)
#logic (necessary?)
flask-login
flask-sqlalchemy
flask-wtf
flask-assets
ddtrace
statsmodels